requirements.txt
tensorflow-cpu
numpy
fastapi
pillow
python-multipart